fix: Remove incorrect int cast for DayOfWeek comparison
Fixed CS0019 error where DayOfWeek enum was being cast to int for comparison.

**Issue**:
WeeklySchedule.DayOfWeek is of type DayOfWeek (enum), not int.
Comparing DayOfWeek with int caused compilation error.

**Fix**:
```csharp
// Before (incorrect):
var todayDayOfWeek = (int)today.DayOfWeek;
ws.DayOfWeek == todayDayOfWeek  // Error: DayOfWeek == int

// After (correct):
var todayDayOfWeek = today.DayOfWeek;
ws.DayOfWeek == todayDayOfWeek  // OK: DayOfWeek == DayOfWeek
```

This allows proper enum-to-enum comparison in EF Core queries.

+++ src/EduPortal.Infrastructure/Services/DashboardService.cs
@@ -85,7 +85,7 @@ public class DashboardService : IDashboardService
                 return ApiResponse<TeacherDashboardStatsDto>.ErrorResponse("Teacher not found");
 
             var today = DateTime.UtcNow.Date;
-            var todayDayOfWeek = (int)today.DayOfWeek;
+            var todayDayOfWeek = today.DayOfWeek;
 
             var stats = new TeacherDashboardStatsDto
             {
